First version of the code. Still needs improvements on the fire spread elipse.

+++ config/landscape/landscape.py
@@ -3,8 +3,8 @@ from dataclasses import dataclass
 
 @dataclass
 class Location:
-    latitude: int
-    longitude: int
+    latitude: float
+    longitude: float
 
 
 @dataclass
